Fix to a bug introduced in previous commit

+++ library/LibUtilities/Memory/ThreadSpecificPool.hpp
@@ -79,7 +79,7 @@ namespace Nektar
         class ThreadSpecificPool
         {
             public:
-                ThreadSpecificPool(unsigned int ByteSize) :
+                ThreadSpecificPool(size_t ByteSize) :
                     m_pool(),
                     m_blockSize(ByteSize),
                     m_mutex()
@@ -134,7 +134,7 @@ namespace Nektar
             private:
                 //boost::thread_specific_ptr<boost::pool<> > m_pool;
                 boost::pool<>* m_pool;
-                unsigned int m_blockSize;
+                size_t m_blockSize;
                 boost::mutex m_mutex;
         };
     }
@@ -142,7 +142,7 @@ namespace Nektar
     class MemPool
     {
         public:
-            typedef std::map<unsigned int, boost::shared_ptr<detail::ThreadSpecificPool> > PoolMapType;
+            typedef std::map<size_t, boost::shared_ptr<detail::ThreadSpecificPool> > PoolMapType;
             
         public:
             MemPool() :
@@ -180,7 +180,7 @@ namespace Nektar
             /// Important: All memory allocated from this method must be returned to the pool
             /// via the Deallocate method.  Deleting pointers allocated from the memory pool with the 
             /// delete operator will result in undefined behavior.
-            void* Allocate(unsigned int bytes)
+            void* Allocate(size_t bytes)
             {
                 if( bytes <= 4 )
                 {
@@ -204,7 +204,7 @@ namespace Nektar
             ///
             /// \attention It is an error to deallocate memory not allocated
             /// from this pool.  Doing this will result in undefined behavior.
-            void Deallocate(void* p, unsigned int bytes)
+            void Deallocate(void* p, size_t bytes)
             {
                 if( bytes <= 4 )
                 {
@@ -226,8 +226,8 @@ namespace Nektar
             
         private:
             detail::ThreadSpecificPool m_fourBytePool;
-            std::map<unsigned int, boost::shared_ptr<detail::ThreadSpecificPool> > m_pools;
-            unsigned int m_upperBound;
+            std::map<size_t, boost::shared_ptr<detail::ThreadSpecificPool> > m_pools;
+            size_t m_upperBound;
     };
 
     LIB_UTILITIES_EXPORT MemPool& GetMemoryPool();
